+++ app/components/ui/index.js
@@ -5,6 +5,7 @@ export { default as ControlButton } from './ControlButton';
 export { default as ActionButton } from './ActionButton';
 export { default as Input } from './Input';
 export { default as Select } from './Select';
+export { default as Checkbox } from './Checkbox';
 export { default as StatusBadge } from './StatusBadge';
 export { default as ModeIndicator } from './ModeIndicator';
 export { default as Pagination } from './Pagination';